% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRS.R
\name{analyzeRS}
\alias{analyzeRS}
\title{Perform response surface analysis}
\usage{
analyzeRS(dat = dat_rs, grp, B.CP = 500, x1 = "d1", x2 = "d2")
}
\description{
Old name of the function, for reference purposes: responseSurface
}
