% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase.R
\name{camelCase}
\alias{camelCase}
\title{Turn strings into CamelCase}
\usage{
camelCase(s, strict = FALSE)
}
\arguments{
\item{s}{a vector of strings}
}
\description{
Often, the variables in data come in different forms. If needed, camel case 
to uniform them.
}
\examples{
camelCase('this_is_fun')
}
