% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence interval}
\usage{
ci(x, myInt = 0.975)
}
\arguments{
\item{x}{A vector}

\item{myInt}{A confidence interval, defaults to 0.975 i.e. 95% confidence
interval}
}
\description{
Find the confidence interval. Missing values are removed by default.
}
\examples{
ci(1:10)
}
