% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumAUC.R
\name{cumAUC}
\alias{cumAUC}
\title{Cumulative Area Under the Curve}
\usage{
cumAUC(x, y)
}
\arguments{
\item{dat}{}
}
\description{
Find cumulative area under the curve
}
\examples{
cumAUC(subset(beaver1, day == 346))
}
