% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitManually.R
\name{fitManually}
\alias{fitManually}
\title{Fit marginal manually}
\usage{
fitManually(grp, grp.val, dat = dat_rs)
}
\description{
More here https://github.com/openanalytics/BIGL/issues/4 
Somehow fitting monotherapies (marginal curves) with drc and feeding into
BIGL::fitMarginals just the problematic coefficient (Emax, BIGL's 'm') works
}
