% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSurf.R
\name{fitSurf}
\alias{fitSurf}
\title{Fit surface model}
\usage{
fitSurf(x, dr, model = "loewe2", B = B.CP)
}
\description{
Convenience function to fit with most common parameters
Skips to next step in case of an error
}
