% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gci.R
\name{gci}
\alias{gci}
\title{Geometric confidence interval}
\usage{
gci(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Find geometric confidence interval as a multiplication factor. Missing values
are removed by default. Currently, defaults to 95% confidence level.
}
\examples{
vec = c(1:10, 100)
gmu(vec); gci(vec)
}
