% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDdiCounts.R
\name{getDdiCounts}
\alias{getDdiCounts}
\title{Get drug-drug interaction counts}
\usage{
getDdiCounts(rs)
}
\arguments{
\item{rs}{a list of a list of response surface objects}
}
\description{
Function extracts the number of synergies and factor encoding to ready for
plotting.
}
