% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmu.R
\name{gmu}
\alias{gmu}
\title{Geometric mean}
\usage{
gmu(x, trim = 0)
}
\arguments{
\item{x}{a vector}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of ‘x’ before the mean is computed. Values of trim outside that range are
taken as the nearest endpoint.}
}
\description{
Find geometric mean. Missing values are removed by default.
}
\examples{
vec = c(1:10, 100)
mean(vec); gmu(vec)
}
