% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd.R
\name{gsd}
\alias{gsd}
\title{Geometric standard deviation}
\usage{
gsd(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Find geometric standard deviation as a multiplication factor. Missing values are removed by default.
}
\examples{
vec = c(1:10, 100)
gmu(vec); gsd(vec)
}
