% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{gse}
\alias{gse}
\title{Geometric standard error}
\usage{
gse(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Find geometric standard error as a multiplication factor. Missing values are removed by default.
}
\examples{
vec = c(1:10, 100)
gmu(vec); gse(vec)
}
