% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseq.R
\name{gseq}
\alias{gseq}
\title{Geometric sequence}
\usage{
gseq(start, end, ratio)
}
\arguments{
\item{start}{the first term in sequence}

\item{end}{the last term in sequence}

\item{ratio}{the common ratio (i.e. the multiple that defines the sequence)}
}
\description{
Create geometric sequence
}
\examples{
gseq(1, 16, 2)
}
