% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsum.R
\name{gsum}
\alias{gsum}
\title{Geometric summaries}
\usage{
gsum(x, dec = 3)
}
\arguments{
\item{x}{A vector}

\item{dec}{Decimal place to round to}
}
\value{
'gsum' returns geometric versions of mean ('gmu'), standard
deviation ('gsd'), standard error ('gse'), and 95\% confidence interval
('gci').
}
\description{
Derive some geometric summaries
}
\details{
Geometric estimates of spread are factors of
multiplication/division.
}
\examples{
gsum(c(1:10, 100))
}
