% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins.R
\name{ins}
\alias{ins}
\title{Insert}
\usage{
ins(vec, element, idx)
}
\arguments{
\item{vec}{a vector}

\item{element}{the element to insert}

\item{idx}{the index at which to insert}
}
\description{
Insert elements into a vector using index
}
\examples{
vec = rep(TRUE, 5) 
ins(vec, FALSE, 2)
ins(vec, FALSE, c(2, 4))
}
