% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOutlier.R
\name{isOutlier}
\alias{isOutlier}
\title{Find outliers}
\usage{
isOutlier(x, method = "mad", alpha = 0.05, pvalue = FALSE, thr = 3, lod = NULL)
}
\arguments{
\item{x}{A vector}

\item{method}{A character, one of 'mad'(default), 'iqr', 'dixon', or 'qn'.}

\item{alpha}{A numeric, significance level for Dixon test, defaults to 0.05.}

\item{pvalue}{A logical, whether to return pvalue for Dixon test.}

\item{thr}{A nonnegative constant that determines the extension of bounds 
(alpha level). Commonly used values are 2, 2.5, 3 (default), and 3.5.}

\item{lod}{A numeric (defaults to NULL), lower boundary for signal that can
reliably distinguished from noise i.e. Limit of Detection. Make sure that if
x is log transformed, you do the same to lod.}
}
\value{
Returns a logical vector by default. If test is Dixon and pvalue is
  set TRUE, returns a list consisting of logical vector and p-value.
}
\description{
Find outliers based on MAD, IQR, Dixon test, or Qn.
}
\examples{
x = c(0.5, 1, 2, 10)
isOutlier(x)
isOutlier(x, 'dixon')
isOutlier(x, 'dixon', 0.1)
isOutlier(x, 'dixon', pvalue = TRUE)

x = c(0.5, 1, 2, 9)
isOutlier(x)            # one outlier
isOutlier(x, 'dixon')   # no outliers

# cautionary example
set.seed(12)  
x = rnorm(4)  # normal distribution
isOutlier(x)                # yet one outlier
isOutlier(x, 'dixon')       # no outliers
isOutlier(x, 'dixon', 0.1)  # one outlier
isOutlier(x, 'dixon', pvalue = TRUE)
}
\references{
Boris Iglewicz and David Hoaglin (1993), Volume 16: How to
Detect and Handle Outliers, The ASQC Basic References in Quality Control:
Statistical Techniques, Edward F. Mykytka, Ph.D., Editor.

Rousseeuw PJ, Croux C. Alternatives to the median absolute deviation. J Amn
Statist Assoc 1993, 88:1273– 1283.
}
