% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIrisFiles.R
\name{loadIrisFiles}
\alias{loadIrisFiles}
\title{Load iris files}
\usage{
loadIrisFiles(folder.path, keep.folder.name = F)
}
\arguments{
\item{folder.path}{Path to folder}

\item{keep.folder.name}{Boolean, defaults to FALSE}
}
\description{
Load all iris files
}
