% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerCaseNamesDT.R
\name{lowerCaseNamesDT}
\alias{lowerCaseNamesDT}
\title{Lower case names in data.table}
\usage{
lowerCaseNamesDT(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
Set names to lowercase in data.table
}
