% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\docType{data}
\name{perimeter}
\alias{perimeter}
\title{384w plate perimeters}
\format{
A data frame with 384 rows and 3 variables:
\tabular{rl}{
 row        \tab A row at 384 well density \cr
 column     \tab A column at 384 well density \cr
 Perimeter  \tab A perimeter at 384 well density \cr
}
}
\usage{
data(perimeter)
}
\description{
A helper datatable to define perimeters on 384w plate
}
\keyword{datasets}
