% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt3D.R
\name{plt3D}
\alias{plt3D}
\title{Plot drug-drug interaction plots in 3D}
\usage{
plt3D(mod, zlim = c(-4.5, 4))
}
\description{
Mostly a convenience function to have the most common settings/appearance for
manuscript: viewangle, consistent colors, range of z-axis etc.
}
