% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltCpd.R
\name{pltCpd}
\alias{pltCpd}
\title{Plot compound}
\usage{
pltCpd(cpd, spc = NULL, DT = dat, legPos = "none", nRow = 3)
}
\description{
Plot individual trace against the backdrop of all others. Highlight DMSO
control
}
