% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltRS.R
\name{pltRS}
\alias{pltRS}
\title{Plot response surface}
\usage{
pltRS(x, color = c("z-score", "maxR", "occupancy", "confint"), ...)
}
\description{
domesticated BIGL::plot.ResponseSurface
note that it calls a subfunction pltResponseSurface (code below)
}
