% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrt.R
\docType{data}
\name{qrt}
\alias{qrt}
\title{Convenience lookup for quadrants}
\format{
A data frame with 1488 rows and 39 variables:
\tabular{rl}{
 Qrt     \tab A quartile at 384 well density \cr
 Row     \tab A row at 384 well density \cr
 Col     \tab A column at 384 well density \cr
}
}
\usage{
data(qrt)
}
\description{
Add quadrants to your 384w experiment using the lookup
}
\keyword{datasets}
