% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleMult.R
\name{scaleMult}
\alias{scaleMult}
\title{Scale multiplicatively}
\usage{
scaleMult(x, target.values)
}
\arguments{
\item{x}{a vector}

\item{target.values}{a vector, can be of length one}
}
\description{
Find geometric mean. Missing values are removed by default.
}
