% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitStr.R
\name{splitStr}
\alias{splitStr}
\title{Split strings}
\usage{
splitStr(string.vector, delimiter, index = NA)
}
\description{
Split strings (or vectors of strings) into vectors or extract a single string
based on index
}
