% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_html.R
\name{theme_html}
\alias{theme_html}
\title{Custom ggplot theme}
\usage{
theme_html(
  base_size = 9,
  base_family = "",
  font_size = 9,
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 18/14
)
}
\description{
Custom theme for ggplot
}
