% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pdf.R
\name{theme_pdf}
\alias{theme_pdf}
\title{Custom ggplot theme}
\usage{
theme_pdf(
  base_size = 18,
  base_family = "",
  font_size = 18,
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 18/14
)
}
\description{
Custom theme for ggplot
}
